no need to move queueBatch function:
